*Actually* support different ODE integrators, fix LF solver
Now separate out the `calcFluxField` function for the fields
as needed to use the new odesolvers and make the program
useable again. Add the solver to the F1DSimulation definition
in the header as a parameter for the c'tor. Add odesolver
files to the project.

Correctly call `calcPhysicalFlux` from within global LF
solvers.